+++ engine/src/conversion/analysis/fun/mod.rs
@@ -14,7 +14,7 @@
 
 mod bridge_name_tracker;
 pub(crate) mod function_wrapper;
-mod implicit_constructors;
+mod implicit_constructor_rules;
 mod overload_tracker;
 mod rust_name_tracker;
 mod subclass;
